+++ lib/date_time_basics.dart
@@ -83,31 +83,6 @@ extension DateTimeBasics on DateTime {
   /// the same contract.
   bool isAtOrAfter(DateTime other) => isAtSameMomentAs(other) || isAfter(other);
 
-  /// Copies a [DateTime], overriding specified values.
-  ///
-  /// A UTC [DateTime] will remain in UTC; a local [DateTime] will remain local.
-  DateTime copyWith({
-    int? year,
-    int? month,
-    int? day,
-    int? hour,
-    int? minute,
-    int? second,
-    int? millisecond,
-    int? microsecond,
-  }) {
-    return (isUtc ? DateTime.utc : DateTime.new)(
-      year ?? this.year,
-      month ?? this.month,
-      day ?? this.day,
-      hour ?? this.hour,
-      minute ?? this.minute,
-      second ?? this.second,
-      millisecond ?? this.millisecond,
-      microsecond ?? this.microsecond,
-    );
-  }
-
   /// Adds a specified number of days to this [DateTime].
   ///
   /// Unlike `DateTime.add(Duration(days: numberOfDays))`, this adds calendar